Share the MCP content vocabulary through a new component
The Agent Client Protocol adopts MCP's JSON representations for content
verbatim, so the two protocols should agree on those types by construction
rather than by a second copy kept in step by hand. The content cluster needs
none of `synesthesia.core`'s HTTP-server closure, so it moves into a component
of its own that depends on `jacinta.core` alone, which a client spawning agents
as subprocesses can depend on without acquiring a web stack.

`object Mcp` re-exports every member, so MCP code still reads `Mcp.TextContent`.
Two adjustments follow from the split: `SamplingMessageContentBlock` becomes a
union alias, because three of its members now precede `core` and a class cannot
extend a trait declared after it — its codecs anchor in `ToolUseContent`'s
companion, which the union's implicit scope includes — and `ImageContent` gains
the optional `uri` ACP defines, which vanishes from the MCP wire when unset.
